Pharos::Logging for use in addons
